+++ crates/topcoat-router/grammar/src/lib.rs
@@ -1,7 +1,6 @@
 #![cfg_attr(docsrs, feature(doc_cfg))]
 
-pub mod error_attr;
-mod handler_args;
+pub mod common;
 pub mod layer;
 pub mod layout;
 pub mod method;
