+++ crates/topcoat-router/grammar/src/route.rs
@@ -8,7 +8,7 @@ use syn::{
 };
 use topcoat_core_grammar::paths::{topcoat_context, topcoat_inventory, topcoat_router};
 
-use super::handler_args::{HandlerArgs, request_ident};
+use super::common::{HandlerArgs, request_ident};
 use super::method::Methods;
 
 pub struct RouteAttr {
